[+] synthetic symbols implementation moved to different files [~] extended API of synthetic symbols: remove by address, remove by mask, remove all, query addres for dbgModuleClass [+] added processing debug event callbacks [+] all synthetic symbols are restored after reload (processing debug event callback)
git-svn-id: https://pykd.svn.codeplex.com/svn@62091 9b283d60-5439-405e-af05-b73fd8c4d996

+++ pykd/dbgtype.cpp
@@ -550,11 +550,19 @@ typedVarClass::printField(const TypeInfo::TypeField &field, stringstream &sstr) 
                 else if ( field.type.name() == "unsigned short*" )
                 {
                     wchar_t   wbuf[0x100];
-                    if ( loadWStrToBuffer( val, wbuf, sizeof(wbuf) ) )      
+                    if ( loadWStrToBuffer( val, wbuf, sizeof(wbuf) ) )
                     {
                         char  mbBuf[0x100] = { 0 };
 
-                        WideCharToMultiByte( CP_ACP, 0, wbuf, wcslen(wbuf)+1, mbBuf, sizeof(mbBuf),  NULL, NULL );
+                        WideCharToMultiByte(
+                            CP_ACP,
+                            0,
+                            wbuf,
+                            (ULONG)wcslen(wbuf) + 1,
+                            mbBuf,
+                            sizeof(mbBuf),
+                            NULL,
+                            NULL);
 
                         sstr << "  (" << mbBuf << " )";
                     }
